More work on external game hosts.
