+++ src/Gameboard.Api.Tests.Integration/Fixtures/Services/TestGamebrainService.cs
@@ -9,6 +9,9 @@ internal class TestGamebrainService : IExternalGameHostService
         return Task.CompletedTask;
     }
 
+    public IQueryable<GetExternalGameHostsResponseHost> GetHosts()
+        => Array.Empty<GetExternalGameHostsResponseHost>().AsQueryable();
+
     public Task<IEnumerable<ExternalGameClientTeamConfig>> StartGame(ExternalGameStartMetaData metaData, CancellationToken cancellationToken)
     {
         return Task.FromResult(Array.Empty<ExternalGameClientTeamConfig>().AsEnumerable());
