+++ src/Gameboard.Api/Features/Player/PlayerService.cs
@@ -691,6 +691,7 @@ public class PlayerService
             string newId = _guids.GetGuid();
             // compute complete score, including bonuses
             var teamScore = await _scores.GetTeamScore(team.Key, CancellationToken.None);
+            var overallScore = (int)Math.Floor(teamScore.OverallScore.TotalScore);
 
             foreach (var player in team)
             {
@@ -704,12 +705,13 @@ public class PlayerService
                     AdvancedFromGameId = player.GameId,
                     AdvancedFromPlayerId = player.Id,
                     AdvancedFromTeamId = player.TeamId,
-                    AdvancedWithScore = model.WithScores ? player.Score : null,
+                    AdvancedWithScore = model.WithScores ? overallScore : null,
                     ApprovedName = player.ApprovedName,
                     Name = player.Name,
                     SponsorId = player.SponsorId,
                     Role = player.Role,
-                    Score = model.WithScores ? (int)Math.Floor(teamScore.OverallScore.TotalScore) : 0
+                    Score = model.WithScores ? overallScore : 0,
+                    WhenCreated = _now.Get()
                 };
 
                 enrollments.Add(newPlayer);
